refactor(db): Change table access in stored procedures to be consistent
Every now and again, we get a deadlock in the database. Whilst these are
expected from time to time we may be able to mitigate it further by making
sure we access our tables in a consistent order. This may not always be
possible but we can do it with these three stored procedures.

Fixes #28