fix(client): Ensure ask-password metrics are only logged once.
`view.context()` was called once per render and once per translation,
causing the ask-password metrics to be logged over and over again.

This fix caches the results of context(), so the function is only called
once per screen render. If the screen must be re-rendered, the cache is
cleared and context() is called again.

fixes #2403
fixes #1523