refactor(client): Extract the ResumeToken into its own model.
A new View mixin is created, the ResumeTokenMixin, that provides
`getStringifiedResumeToken` to get a stringified ResumeToken that
can be sent along in emails.

All fxa-client functions that created a resume token from the relier
now accept an options block that can have a `resume` field.

The User and Account models were updated to pass along the stringified
ResumeToken.

issue #2563