fix(client): Fix the submit button on communcation preferences page.
The problem is a view was only allowed to have a single progress indicator,
attached to a single button. If a view was re-rendered with a new submit
button, the new submit button could not have a progress indicator.

Instead of saving a reference to the progress indicator on the view,
save it on the button the behavior is attached to. If the button goes
away, so does the progress indicator, and a new one will be created
when needed.

fixes #2502