fix(client): Ensure status messages are shown on the signup page.
Done using better data sandboxing with View models!

`model` is used for current view data. A View's Model data is created and
passed in to the View, based on data sent from the previous View.

`ephemeralMessages` is no more, it was global mutable state. Terrible.

view.ephemeralData is no more, since it was based on `ephemeralMessages`.
Use the `model` instead.