fix(metrics): Reduce the number of localStorage errors in Sentry.
crosstab writes to localStorage every 3 seconds to find out if any
other tabs are open. Errors are reported to Sentry on every write
if localStorage data is unable to be written to disk.

Inter-tab communication is done via BroadcastChannels if the
browser has support and falls back to crosstab if not. If
the browser supports BroadcastChannels, abort crosstab initialization

issue #3414