fix(client): Reduce the number of `parseMessage` errors.
*Really* ignore messages from ignored origins.

In 7010bbd9 we intended to ignore postMessage messages from the browser. Due to
an error in both the tests and the code, we logged the error to the console
and did *not* ignore the postMessage message.

Now we log the error to the console AND ignore, for real this time.

fixes #3594