feat(server): Enable CSP! (#3627)
* feat(server): Enable CSP!

Both CSP and CSP-report-only can be used at the same time. This allows us
to enable CSP universally, still have the functional tests run, and still see
if anybody is adding resources we do not expect.

Two CSP middlware configurations are created, one that is "blocking", one
"reportOnly". All the former rules are added to "blocking", scriptSrc was
updated to allow "'unsafe-eval'" to allow the functional tests to run. A second
scriptSrc was added to the reportOnly middlware that does not allow
"'unsafe-eval'", this will allow us to see if anybody has been able to cause
an eval to take place.

fixes #2155

* refactor(csp): Extract CSP rules into rules files, use a common CSP middleware