refactor(tests): Use a helper to get the `remote` reference. (#3808)
* refactor(tests): Use a helper to get the `remote` reference.


r=@philbooth

Thanks for the suggestions Phil!
This allows us to pass `this.parent` from within `then` callbacks
to functions that need a remote reference.

* chore(tests): Updates based on @philbooth's feedback.

Instead of using ternary operators, use logical ORs to decide
which `remote` to return. This allows us to add `parent` to the
mix as well! Updated tests that pass `this.parent` to `openPage`,
we've got this taken care of now.