fix(client): Fix sign in from the firstrun flow. (#3889)
* fix(client): Fix sign in from the firstrun flow.

Two `fxaccounts:login` messages were being sent, both with the same
sessionToken and keyFetchToken. When the 2nd message was received by
the browser, it would attempt to re-use the keyFetchToken to fetch
keys. Since keyFetchTokens are single use, this failed.

This occurred because the signin confirmation flow called the
`beforeSignUpConfirmationPoll` broker method before the confirmation poll,
which sent one notice, and then `afterSignIn` when the poll completed, which
sent another notice.

The solution is to only send one `fxaccounts:login` message.

Instead of re-using `afterSignIn`, a new broker method is created,
`afterSignInConfirmationPoll`. The firstrun flow uses
`afterSignInConfirmationPoll` to notify the firstrun page of the signin, but
does NOT notify the browser.

fixes #3880

r=@vbudhram