fix(notifications): Allow a notification handler to be spied upon. (#4732) r=@philbooth
While adding handlers for items in `notifications`, if the handler
was declared using a string, the actual handler method reference was
immediately created. This made it impossible to spy upon a method
declared this way.

The fix is to do "late-binding" where the handler method is looked up
when the notification is triggered. This allows sinon to set up a spy
and then have the spy called whenever the notification is triggered.

fixes #4731