refactor(settings): Simplify settings header rendering. (#4744) r=@philbooth
What was the problem?
The settings header format was duplicated in both the template
and in code. In addition, updating the settings header was
done imperatively, manually after leaving a child view.
This just felt a bit complex and gross. Not very MVC.

How does this fix that?
Extract the settings header HTML into a template. Render the
template on view render and on account change:email or
change:displayName.