feat(tests): Reload mocha tests whenever source/js changes. (#4752) r=vladikoff,vbudhram
What is the problem?
To run the unit tests while developing, the dev has to leave
their editor, go to the browser tab with the unit tests open,
and refresh the page. That's a big context shift.

How does this fix it?
Add the `grunt watch:livereload` task which forces the mocha
tests to reload any time a test or source file changes.

Do the tests still run w/o the grunt task running?
Yes, but they will not automatically reload whenever
source files change.