fix(metrics): Fix the "refreshes_metrics" test. (#4850) r=@philbooth, @vladikoff
What is the problem?
The `_send` function on Metrics.prototype was always being called instead of
the variant on StorageMetrics.prototype.

This happened because Metrics.initialize was called as part of the constructor when setting
up the prototype chain. Metrics.initialize creates `this._flush` by binding `this.flush`
to the current context, which is a metrics instance. When `this._flush` was called,
the context was the original metrics object. `this._flush` would call `this._send` on
the original object too, because of the context.

How does this fix it?
Instead of calling `new Metrics()` to set up the prototype chain, copy
over the functions from Metrics' prototype to StorageMetrics' prototype,
then from the StorageMetrics constructor, call the Metrics constructor. It
sets up the prototype chain correctly, and allows functions to bind correctly.

Why remove the Notifier mock?
App-start passes StorageMetrics a notifier instance, which is then passed
to the Metrics constructor, no need for a mock.

fixes #4844