fix(test): Fix "verify different browser, verification_redirect=always" test. (#4869) r=vladikoff
* fix(test): Fix "verify different browser, verification_redirect=always" test.

What is the problem?
We often see a StaleElementReference error when getting the body text.
This can happen whenever `findByCssSelector('body')` is called when FxA
is present and `getVisibleText` is called when 123done loads.

How does this fix it?
Look at the URL, if the URL is the OAuth app, then try to get the body.

issue #4847

* fix(test): Fix the "original tab closed, verification_redirect=always" test

What is the problem?
Timing issues! There is a lot of redirection going on, and we aren't sure
what page the user will land on. We were making comparisons on stale elements,
and when we haven't reached a steady state.

How does this fix it?
It's gross. But we wait 5 seconds. It's not failsafe, but it's better.
Then, compare the current URL with the two possible URLs. Then compare
screen state based on the current URL.

issue #4847