fix(test): Fix NoSuchBrowserNotification flakiness. (#4883) r=@jrgm
Tests would often fail with a NoSuchNotification message.
If notifications are checked for before the screen
transitions, it's possible the XHR request was still
outstanding and the web channel message isn't 
yet sent. WebChannel notifications must be 
received within 5 seconds, otherwise the test fails. 
We give screen transitions up to 20 seconds, because 
transitions can be very slow when run against real servers.

This changes all checks for browser notifications to occur after
a screen transition.

The only ones that haven't been updated are the "change password"
tests in *_settings.js, I haven't put in the effort to figure out
the proper way of doing that yet.

fixes #4882