fix(sms): Fix the SMS experiment bucketing logic. (#4977) r=vladikoff
Issue #4944


Users in the SMS "control" group must be in CAD, we are comparing
SMS against the CAD screens, not the general population.

This PR makes that happen. The SMS/CAD selection logic is extracted
into its own mixin so that it can be independently tested and
used on the "signup" page too.

From the experiment-mixin, trigger `flow.initialized` which
will force the flow model to initialize for experiments
that are created at runtime.