refactor(build): Replace babel-preset-2015-nostrict with babel-preset-2015 (#5161) r=vladikoff
babel-preset-2015 is maintained by the babel team, the -nostrict variant
is not. I used the -nostrict variant because it doesn't try to wrap
all of our modules in an AMD wrapper, but I didn't know it is
possible to supply the "modules: false" option to the plugin
to have the same effect.