refactor(mixins): Mixins can now declare other mixins as dependencies. (#5251) r=@shane-tomlinson
What is the problem?
A lot of mixins require other mixins to function. To make this work,
a View would have to mix-in all of the dependent mixins too, or
a mixin would have to mix the dependent mixin to itself. It was
difficult to know what to do whenever adding a mixin that depended
on others, like the connect-another-device-mixin.

How does this fix the problem?
A mixin that depends on other mixins can declare an array `dependsOn`.
Each mixin declared in `dependsOn` will be mixed into the
consuming object before the mixin itself.

This also takes care of the possibility of a mixin being added
multiple times.