refactor(service): remove smtp settings (#70) r=@vladikoff
Fixes #61 .

Not much to talk about here, just deleted the settings and adjusted the tests.

The Cargo.lock file was updated as well, I think that's because some of the crates were updated. Since we got >= in the cargo versions right now, I think when crates are updated we just grab the new versions on cargo build, thus updating the Cargo.lock file.