fix(email): gracefully handle errors from fxa-email-service (#2510) r=@vladikoff
If, say, `fxa-email-service` isn't running and you try to send email
using it, the request fails and no `body` is passed to the response
handler. This causes the auth server to fail with a 500.

Better test coverage would have picked this up, so I've opened #2509 to
address that. In the meantime, this is a quick hack if we want to patch
it in train 116.