feat(email): add a service property to the X-SES-MESSAGE-TAGS header
Adding a `service` property to the `X-SES-MESSAGE-TAGS` header lets us
track the email-sending service in AWS metrics. If something goes wrong
with the new service, this will help ops to identify the problem.

While implementing that, I also opted to rejig some of the surrounding
code in `lib/senders/email.js`, which was unnecessarily repetitive in
some parts and difficult to follow in others. I think far greater
improvements can still be made in that module, though, if we pursue a
more data-driven/functional approach to the whole thing.

https://github.com/mozilla/fxa-auth-server/pull/2526
r=vbudhram