fix(metrics): use correct format for email service notifications
The email service made a conscious break with the SQS notification
structure, to make the recipients array more uniform. Rather than
sometimes being an object with an `emailAddress` property and other
times being a string, it makes the items plain strings on all
notification types.

When I added code for the new notification queue to this repo, I forgot
about that and just ported the SQS logic verbatim. This fixes it to
always expect plain strings in the recipient arrays.