feat(scripts): add ROW_COUNT() checks to the procedure-linting script
In https://bugzilla.mozilla.org/show_bug.cgi?id=1499819, it was found
that `ROW_COUNT()` causes db replication to lag. As such, it makes sense
to grep for `ROW_COUNT()` in our stored procedure lint script.

The script was called `explain-warn.js` before, but in light of the new
functionality it's renamed here to the more generic `procedure-lint.js`.