chore(scripts): lint-ignore consumeRecoveryCode_2 and setPrimaryEmail_3
These two scripts use ROW_COUNT() but I couldn't see an obvious way to
work round them. In the case of consumeRecoveryCode_2, ROW_COUNT() is
used as an interim signal for aborting the remainder of the procedure.
In the case of setPrimaryEmail_3, the UPDATE is in a transaction which
causes affectedRows to be zero upon returning.