fix(scripts): stop the explain script tripping over git grep colours
The explain script was using `cut` to split paths from git grep before
the colon character. This fails on system where control characters such
as colours are emitted before the colon.

A more robust approach is to use a regex to pull out the specific text
we're looking for. The script is already hard-coded with other paths
like `lib/db/mysql.js`, so matching against the expected migration path
doesn't seem like too much of a stretch to me.