chore(db): use mariadb-friendly drop index syntax
Sometimes we get self-hosters or contributors showing up who want to use
MariaDB. Even though the official line is that we don't support it, it
happens often enough that we may as well just fix the places where we're
incompatible. Afaik, that is just the `DROP INDEX` syntax in a couple of
places.

Note there are also occurrences in the reverse migrations, which I
haven't bothered to fix. Seeing as they're not invoked in normal running
and need to be manually uncommented before running anyway, it seems
reasonable for them to be manually edited if anyone needs to do so.